Fix for playing on device
Specify that the format used for all the AVAudioEngine node
connections is the format of the PCM buffer.

This fixes a crash when playing recorded audio on a real device.
The issue was that the recorded audio was in mono (single channel),
but the other nodes were defaulting to stereo (two channels).

Apparently something in the AVFoundation libraries changed,
because it worked without this change just two or three
months ago.